+++ src/notifications/ManagerLinux.cpp
@@ -27,7 +27,7 @@
 #include "MxcImageProvider.h"
 #include "UserSettingsPage.h"
 #include "Utils.h"
-#include "timeline/NhekoDBusInterface.h"
+#include "NhekoDBusInterface.h"
 
 NotificationsManager::NotificationsManager(QObject *parent)
   : QObject(parent)
